+++ tests/testthat/test-sensitivity-tornado-plot.R
@@ -36,7 +36,12 @@ results <- sensitivityCalculation(
 test_that("sensitivityTornadoPlot fails with incorrect input", {
   expect_error(
     sensitivityTornadoPlot("x"),
-    "argument 'sensitivityCalculation' is of type 'character', but expected 'SensitivityCalculation'"
+    regexp = messages$errorWrongType(
+      "sensitivityCalculation",
+      "character",
+      "SensitivityCalculation"
+    ),
+    fixed = TRUE
   )
 })
 
@@ -44,6 +49,15 @@ test_that("sensitivityTornadoPlot fails with invalid parameterFactor", {
   expect_error(
     sensitivityTornadoPlot(results, parameterFactor = 0),
     "parameterFactor error"
+  )
+})
+
+test_that("sensitivityTornadoPlot fails with invalid xAxisZoomRange", {
+  xAxisZoomRange <- 100
+  expect_error(
+    sensitivityTornadoPlot(results, xAxisZoomRange = xAxisZoomRange),
+    messages$errorWrongLength(xAxisZoomRange, 2),
+    fixed = TRUE
   )
 })
 
@@ -117,6 +131,23 @@ test_that("sensitivityTornadoPlot works with custom PK parameter", {
     fig = suppressWarnings(p)
   )
 })
+
+
+# Default plot with x-axis zoom -------------------------------------------
+
+test_that("sensitivityTornadoPlot applies x-axis zoom range correctly", {
+  set.seed(123)
+  p <- sensitivityTornadoPlot(results, xAxisZoomRange = c(-100, 100))
+
+  set.seed(123)
+  suppressWarnings(
+    vdiffr::expect_doppelganger(
+      title = "sensitivityTornadoPlot zoomed",
+      fig = p
+    )
+  )
+})
+
 
 # Multiple output paths ---------------------------------------------------