+++ tests/testthat/test-utilities-data.R
@@ -215,7 +215,8 @@ test_that("It throws an error when molWeights of data sets are different and no 
   dataSet2$molWeight <- 2
   expect_error(
     calculateMeanDataSet(list(dataSet1, dataSet2)),
-    messages$errorOutputMolWeightNeeded()
+    regexp = messages$errorOutputMolWeightNeeded(),
+    fixed = TRUE
   )
 })